(feat) cast to int the start and end time

+++ core/data_sources/trades_feed/trades_feed_base.py
@@ -13,12 +13,12 @@ class TradesFeedBase(ABC):
     def get_exchange_trading_pair(self, trading_pair: str) -> str:
         ...
 
-    async def get_historical_trades(self, trading_pair: str, start_time: float, end_time: Optional[float] = None):
+    async def get_historical_trades(self, trading_pair: str, start_time: int, end_time: Optional[int] = None):
         if not end_time:
             end_time = time.time()
         historical_trades = await self._get_historical_trades(trading_pair, start_time, end_time)
         return historical_trades
 
     @abstractmethod
-    async def _get_historical_trades(self, trading_pair: str, start_time: float, end_time: float):
+    async def _get_historical_trades(self, trading_pair: str, start_time: int, end_time: int):
         ...
